+++ src/helper/retryRequester.ts
@@ -15,16 +15,23 @@
  * limitations under the License.
  */
 
-import * as Q from 'q';
+import * as Promise from 'any-promise';
+import { PlywoodRequester, DatabaseRequest } from 'plywood-base-api';
 
 export interface RetryRequesterParameters<T> {
-  requester: Requester.PlywoodRequester<T>;
+  requester: PlywoodRequester<T>;
   delay?: number;
   retry?: int;
   retryOnTimeout?: boolean;
 }
 
-export function retryRequesterFactory<T>(parameters: RetryRequesterParameters<T>): Requester.PlywoodRequester<T> {
+function delayPromise(n: number): Promise<void> {
+  return new Promise<void>((resolve) => {
+    setTimeout(resolve, n);
+  });
+}
+
+export function retryRequesterFactory<T>(parameters: RetryRequesterParameters<T>): PlywoodRequester<T> {
   let requester = parameters.requester;
   let delay = parameters.delay || 500;
   let retry = parameters.retry || 3;
@@ -33,14 +40,14 @@ export function retryRequesterFactory<T>(parameters: RetryRequesterParameters<T>
   if (typeof delay !== "number") throw new TypeError("delay should be a number");
   if (typeof retry !== "number") throw new TypeError("retry should be a number");
 
-  return (request: Requester.DatabaseRequest<T>): Q.Promise<any> => {
+  return (request: DatabaseRequest<T>): Promise<any> => {
     let tries = 1;
 
-    function handleError(err: Error): Q.Promise<any> {
+    function handleError(err: Error): Promise<any> {
       if (tries > retry) throw err;
       tries++;
       if (err.message === "timeout" && !retryOnTimeout) throw err;
-      return Q.delay(delay).then(() => requester(request)).catch(handleError);
+      return delayPromise(delay).then(() => requester(request)).catch(handleError);
     }
 
     return requester(request).catch(handleError);